refactor: 3-layer 전환 1단계 — ingest 모듈 Port/Adapter 제거
