+++ src/main/java/com/bestduo_BE/common/infra/persistence/projection/MatchQueueStatusCount.java
@@ -1,4 +1,4 @@
-package com.bestduo_BE.ingest.application.port;
+package com.bestduo_BE.common.infra.persistence.projection;
 
 public interface MatchQueueStatusCount {
   String getStatus();
