+++ src/main/java/com/bestduo_BE/ingest/application/IngestMatchDetail.java
@@ -1,8 +1,8 @@
 package com.bestduo_BE.ingest.application;
 
 import com.bestduo_BE.common.application.port.RiotApiPort;
-import com.bestduo_BE.ingest.application.port.BottomDuoRawSaver;
-import com.bestduo_BE.ingest.application.port.MatchSaver;
+import com.bestduo_BE.ingest.infra.persistence.BottomDuoRawSaver;
+import com.bestduo_BE.ingest.infra.persistence.MatchSaver;
 import com.bestduo_BE.common.domain.model.BottomDuoRaw;
 import com.bestduo_BE.common.domain.model.IngestResult;
 import com.bestduo_BE.common.domain.model.Tier;
