--- src/main/java/com/bestduo_BE/ingest/application/port/MatchSaver.java
@@ -1,7 +1,0 @@
-package com.bestduo_BE.ingest.application.port;
-
-import com.bestduo_BE.common.infra.riot.dto.RiotMatchDto;
-
-public interface MatchSaver {
-  void save(String matchId, RiotMatchDto matchDetail);
-}
