+++ src/main/java/com/bestduo_BE/ingest/infra/persistence/MatchSaver.java
@@ -1,22 +1,20 @@
 package com.bestduo_BE.ingest.infra.persistence;
 
-import com.bestduo_BE.ingest.application.port.MatchSaver;
 import com.bestduo_BE.common.infra.persistence.entity.Match;
 import com.bestduo_BE.common.infra.persistence.repository.MatchJpaRepository;
 import com.bestduo_BE.common.infra.riot.dto.InfoDto;
 import com.bestduo_BE.common.infra.riot.dto.RiotMatchDto;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import tools.jackson.databind.ObjectMapper;
 
-@Component
+@Service
 @RequiredArgsConstructor
-public class MatchSaverImpl implements MatchSaver {
+public class MatchSaver {
 
   private final MatchJpaRepository matchRepository;
   private final ObjectMapper objectMapper;
 
-  @Override
   public void save(String matchId, RiotMatchDto matchDetail) {
     InfoDto info = matchDetail.info();
