+++ src/test/java/com/bestduo_BE/ingest/application/IngestMatchDetailPatchFilterTest.java
@@ -16,8 +16,8 @@ import com.bestduo_BE.common.infra.riot.dto.ParticipantDto;
 import com.bestduo_BE.common.infra.riot.dto.RiotMatchDto;
 import com.bestduo_BE.common.infra.riot.dto.TeamDto;
 import com.bestduo_BE.common.application.port.RiotApiPort;
-import com.bestduo_BE.ingest.application.port.BottomDuoRawSaver;
-import com.bestduo_BE.ingest.application.port.MatchSaver;
+import com.bestduo_BE.ingest.infra.persistence.BottomDuoRawSaver;
+import com.bestduo_BE.ingest.infra.persistence.MatchSaver;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
